Fix width for long torrent names
The word-break directive makes a long string break to wrap even in a
non-breakable place in order to stay with in the width of it's parent
element.  This makes it so that there is never any horizontal scrolling
required because the torrent names etc. just wrap to the next line.