Added filter options
This adds all the filter options that the desktop interface has,
including Active, Inactive and Error.  All the statuses have been
combined into a dropdown.

This also adds trackers filter.  It will produce the same list in a
dropdown that the trackerslabel plugin makes, but that plugin is not
required for this to work.